+++ src/r_compliance.cc
@@ -4,13 +4,10 @@
 #include <string.h>
 #include <Rcpp.h>
 #include <ostream>
+#include <RcppThread.h>
 
 void exit_fasttext(int status_code) {
   if (status_code != EXIT_SUCCESS) {
     Rcpp::stop("Failure in fastrtext. Exit code: " + std::to_string(status_code));
   }
-}
-
-namespace std {
-  std::ostream Rcout(Rcpp::Rcout.rdbuf());
 }